git: use curated author information resolved by unix username
This should be a convenient way to automatically retrieve the author
informations for the state repository changes without the need of each
individual user to setup yet another author file.
This is a central service, hence we can easily deploy and generate the
authors file alongside the actual unix user and group permissions rolled
out by ansible.