git: use KEYRING config variable to export GNUPGHOME
This way we have better control and isolation of the keyring between
dbscripts internal sub command env and test suite gpg handling. Export
the GNUPGHOME variable inside the db-functions prolog to ensure any gpg
or git operations requiring a keyring operate on the one defined in the
config. However doing it this way will not influence any custom calls in
the test suite except if explicitly desired.
Furthermore remove all explicit passes of GNUPGHOME to subcommands.