fix(db-functions): do proper error handling for local variables
This paradigm seems to be in dbscripts since a decade, which effectively
is void. Using a subshell with a local variable declaration basically
always leads to success as the local declaration masks the return code
of the subshell, which means an early exit using `|| return 1` will
never work.

Fix this by splitting local variable declaration from assignment.

Signed-off-by: Levente Polyak <anthraxx@archlinux.org>